refactor: 어드민 관련 페이지를 응답하는 메서드를 한 곳으로 이동
- 기존의 AdminController 클래스명을 AdminPageController 로 수정
