+++ src/test/java/roomescape/controller/AdminPageControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class AdminControllerTest {
+public class AdminPageControllerTest {
 
     @LocalServerPort
     int port;
